add helper function MakeGenerateShapeOpAttribute
